first-pass on splitting features,
