+++ crates/bevy_entropy/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bevy_entropy"
-version = "0.9.0"
+version = "0.11.0-dev"
 edition = "2021"
 description = "Bevy Engine's RNG integration"
 homepage = "https://bevyengine.org"
@@ -9,15 +9,21 @@ license = "MIT OR Apache-2.0"
 keywords = ["game", "bevy", "rand", "rng"]
 categories = ["game-engines", "algorithms"]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+bevy_reflect = ["dep:bevy_reflect", "bevy_app/bevy_reflect", "serialize"]
+default = ["bevy_reflect"]
+serialize = ["dep:serde", "rand_core/serde1", "rand_chacha/serde1"]
 
 [dependencies]
-bevy_app = { path = "../bevy_app", version = "0.9.0" }
-bevy_ecs = { path = "../bevy_ecs", version = "0.9.0" }
-bevy_reflect = { path = "../bevy_reflect", version = "0.9.0" }
-serde = { version = "1.0", features = ["derive"] }
-rand_core = { version = "0.6", features = ["std", "serde1"] }
-rand_chacha = { version = "0.3", features = ["serde1"] }
+# bevy
+bevy_app = { path = "../bevy_app", version = "0.9.0", default-features = false }
+bevy_ecs = { path = "../bevy_ecs", version = "0.9.0", default-features = false }
+bevy_reflect = { path = "../bevy_reflect", version = "0.9.0", optional = true }
+
+# others
+serde = { version = "1.0", features = ["derive"], optional = true }
+rand_core = { version = "0.6", features = ["std"] }
+rand_chacha = "0.3"
 
 [dev-dependencies]
 rand = "0.8"
