+++ crates/bevy_entropy/src/prelude.rs
@@ -1,3 +1,3 @@
 pub use crate::component::EntropyComponent;
 pub use crate::resource::GlobalEntropy;
-pub use crate::EntropyPlugin;
+pub use crate::plugin::EntropyPlugin;
